Some fixes (SSE disabled on windows) Some *fixes* VS2013 compilation.. #Not.

+++ include/map/pathfind/Heuristic.h
@@ -10,22 +10,22 @@ namespace Seed {
 class SEED_CORE_API Heuristic
 {
 	public:
-		static f32 Manhattan(u32 x, u32 y)
+		static f32 Manhattan(f32 x, f32 y)
 		{
 			return x + y;
 		}
 
-		static f32 SquareEuclidean(u32 x, u32 y)
+		static f32 SquareEuclidean(f32 x, f32 y)
 		{
 			return (x * x + y * y);
 		}
 
-		static f32 Euclidean(u32 x, u32 y)
+		static f32 Euclidean(f32 x, f32 y)
 		{
 			return sqrt(x * x + y * y);
 		}
 
-		static f32 Chebyshev(u32 x, u32 y)
+		static f32 Chebyshev(f32 x, f32 y)
 		{
 			return std::max(x, y);
 		}
